pyproject.toml
[tool.poetry]
name = "ai-companion-backend"
version = "0.1.0"
description = ""
authors = ["Tak-min <taku888min@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pydantic = "^2.7.1"
pydantic-settings = "^2.3.4"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
google-generativeai = "^0.5.4"
assemblyai = "^0.23.0"
websockets = "^12.0"
# For Auth0
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
fastapi-auth0 = "^0.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
